requirements.txt
fastapi==0.85.1
uvicorn[standard]==0.19.0
pydantic==1.10.2
py3-tts-wrapper[google,microsoft,sapi,sherpaonnx,elevenlabs,watson,polly,witai,playht,espeak,controllable]
fuzzysearch